% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ugh.R
\name{ugh}
\alias{ugh}
\title{Print a Random Inspirational Quote}
\usage{
ugh()
}
\description{
The `ugh()` function prints a random inspirational quote to lift your spirits!
It reads quotes from a JSON file included in the package.
}
